requirements.txt
aws-cdk-lib
aws-sam-cli
boto3
aws-solutions-constructs.core